added comments to FileToSchemaConverter. Updated todo
