+++ src/functions/getFunctionResult.ts
@@ -1,10 +1,14 @@
 /**
- * Get result of a function call.
- * @param args {any} arguments of function. If this is an array, then each array value is passed as a parameter.
+ * Get result of a function call. This is used to quickly call a function with dynamic parameters that could be either an 
+ * array or a single variable such as an object.
+ * 
+ * @param { Function } fn - Function to call.
+ * @param {any} [args] - arguments of function. If this is an array, then each array value is passed as a parameter.
  * Otherwise the whole variable is passed as a single param. If this parameter is undefined, it is not passed.
- * @param fn Function to call.
+ * 
+ * @returns { any } - result of called function
  */
- export default function getFunctionResult( args, fn ) {
+ export default function getFunctionResult(  fn: Function, args?: any ) {
 
     let fn_result = null;
     if ( args && Array.isArray( args ) ) {
