+++ src/types/IFileIO.ts
@@ -1,0 +1,8 @@
+/**
+ * Interface for FileIO constructor. See constructors/FileIO for documentation on each method.
+ */
+export default interface IFileIO {
+    replaceInFile: Function,
+    readDir: Function,
+    readFile: ( file_path: string ) => string
+}
