+++ src/types/MiscTypes.ts
@@ -1,0 +1,12 @@
+/**
+ * Misc types
+ */
+
+/**
+ * Generic object type. Just an object with any number of properties of any value
+ */
+type GenericObject = {
+    [key: string]: any
+}
+
+export { GenericObject };
